+++ src/http_plus_purplelemons_dev/server.py
@@ -14,7 +14,12 @@ if __name__=="__main__":
     parser.add_argument("-p", "--port", type=int, default=8080, help="The port to listen on.")
     parser.add_argument("-d", "--debug", action="store_true", help="Enables debug mode.")
     parser.add_argument("--bind", type=str, default="0.0.0.0", help="The host IP to listen on.")
+    parser.add_argument("---init", action="store_true", help="Does not start the server, but instead initializes the current directory for HTTP+")
     args = parser.parse_args()
+
+    if args.init:
+        http_plus.init()
+        exit(0)
 
     server = http_plus.Server("0.0.0.0", args.port, debug=args.debug)